remove wallet ready screen too
