+++ mobile-app/lib/v2/screens/welcome/welcome_screen.dart
@@ -1,18 +1,84 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:quantus_sdk/quantus_sdk.dart';
+import 'package:resonance_network_wallet/providers/account_providers.dart';
 import 'package:resonance_network_wallet/providers/l10n_provider.dart';
+import 'package:resonance_network_wallet/providers/remote_config_provider.dart';
+import 'package:resonance_network_wallet/services/firebase_messaging_service.dart';
+import 'package:resonance_network_wallet/services/wallet_creation_service.dart';
+import 'package:resonance_network_wallet/shared/extensions/toaster_extensions.dart';
 import 'package:resonance_network_wallet/v2/components/quantus_button.dart';
 import 'package:resonance_network_wallet/v2/components/scaffold_base.dart';
-import 'package:resonance_network_wallet/v2/screens/create/wallet_ready_screen.dart';
+import 'package:resonance_network_wallet/v2/screens/accounts/account_ready_screen.dart';
 import 'package:resonance_network_wallet/v2/screens/import/import_wallet_screen.dart';
 import 'package:resonance_network_wallet/v2/screens/welcome/onboarding_background.dart';
 import 'package:resonance_network_wallet/v2/theme/app_text_styles.dart';
 
-class WelcomeScreenV2 extends ConsumerWidget {
+class WelcomeScreenV2 extends ConsumerStatefulWidget {
   const WelcomeScreenV2({super.key});
 
   @override
-  Widget build(BuildContext context, WidgetRef ref) {
+  ConsumerState<WelcomeScreenV2> createState() => _WelcomeScreenV2State();
+}
+
+class _WelcomeScreenV2State extends ConsumerState<WelcomeScreenV2> {
+  final WalletCreationService _walletCreationService = WalletCreationService();
+
+  static const _accountName = 'Account 1';
+  static const _walletIndex = 0;
+
+  bool _isCreating = false;
+
+  Future<void> _createWallet() async {
+    setState(() => _isCreating = true);
+    try {
+      final mnemonic = await SubstrateService().generateMnemonic();
+      if (mnemonic.isEmpty) throw Exception('Mnemonic generation returned empty.');
+
+      final address = HdWalletService().keyPairAtIndex(mnemonic, 0).ss58Address;
+      final checksum = await HumanReadableChecksumService().getHumanReadableName(address);
+
+      final accounts = ref.read(accountsProvider).value ?? <Account>[];
+      await _walletCreationService.createNewWallet(
+        name: _accountName,
+        mnemonic: mnemonic,
+        walletIndex: _walletIndex,
+        accountId: address,
+        existingAccounts: accounts,
+      );
+
+      ref.invalidate(accountsProvider);
+      ref.invalidate(activeAccountProvider);
+
+      if (ref.read(remoteConfigProvider).enableRemoteNotifications) {
+        ref.read(firebaseMessagingServiceProvider).registerDeviceIfPossible();
+      }
+
+      if (!mounted) return;
+      Navigator.pushAndRemoveUntil(
+        context,
+        MaterialPageRoute(
+          builder: (_) => AccountReadyScreen(
+            accountId: address,
+            accountName: _accountName,
+            checksumPhrase: checksum,
+            origin: AccountReadyOverviewOrigin.walletCreated,
+          ),
+        ),
+        (route) => false,
+      );
+    } catch (e) {
+      if (mounted) {
+        final l10n = ref.read(l10nProvider);
+        context.showErrorToaster(message: l10n.createWalletRecoveryPhraseSaveError(e.toString()));
+      }
+    } finally {
+      if (mounted) setState(() => _isCreating = false);
+    }
+  }
+
+  @override
+  Widget build(BuildContext context) {
     final l10n = ref.watch(l10nProvider);
 
     return ScaffoldBase(
@@ -27,16 +93,7 @@ class WelcomeScreenV2 extends ConsumerWidget {
             child: Text(l10n.welcomeTagline, textAlign: TextAlign.center, style: context.themeText.mediumTitle),
           ),
           const SizedBox(height: 56),
-          QuantusButton.simple(
-            label: l10n.welcomeCreateNewWallet,
-            onTap: () => Navigator.push(
-              context,
-              MaterialPageRoute(
-                settings: const RouteSettings(name: 'create_wallet'),
-                builder: (_) => const WalletReadyScreenV2(),
-              ),
-            ),
-          ),
+          QuantusButton.simple(label: l10n.welcomeCreateNewWallet, onTap: _createWallet, isLoading: _isCreating),
           const SizedBox(height: 24),
           QuantusButton.simple(
             label: l10n.welcomeImportWallet,
@@ -48,6 +105,7 @@ class WelcomeScreenV2 extends ConsumerWidget {
               ),
             ),
             variant: ButtonVariant.secondary,
+            isDisabled: _isCreating,
           ),
           const SizedBox(height: 40),
         ],
